etc/homework/250808/pjt-02/skeleton/problem_C1.py
import os
import requests
import json
from pathlib import Path
from dotenv import load_dotenv

# 1. [ 환경 변수 로드 ]


# 2. [ 최대 100개까지 주제별 도서 데이터를 가져오는 함수 정의 ]
def fetch_books_by_topic(topic, max_results=100):
    pass


# 3. '인공지능' 도서 데이터를 처리하는 함수 정의
def process_ai_books():
    # 3.1 [ '인공지능' 관련 도서 검색 ]
    # fetch_books_by_topic()을 호출하여 '인공지능' 관련 도서를 100개 수집합니다.
    pass  

    # 3.2 [ 수집된 데이터에서 가격 정보가 있는 책 필터링 및 가격순 정렬 ]
    pass 

    # 3.3 [ 상위 10개 도서 선택 ]
    pass 

    # 3.4 [ 상위 10개 도서 정보 출력 ]
    pass

    # 3.5 [ JSON 파일로 저장할 데이터 준비 ]
    # output/ai_top10_books.json 파일로 저장
    pass

    # 3.6 [ 완료 메시지를 출력 ]
    pass

# 함수 실행
if __name__ == '__main__':
    process_ai_books()
